refactor usage of updateTransactions
