+++ src/fiatExchanges/saga.test.ts
@@ -16,7 +16,7 @@ import { Screens } from 'src/navigator/Screens'
 import { AddressRecipient, RecipientType } from 'src/recipients/recipient'
 import { sendPayment, sendPaymentFailure, sendPaymentSuccess } from 'src/send/actions'
 import { tokensByCurrencySelector } from 'src/tokens/selectors'
-import { updateTransactions } from 'src/transactions/actions'
+import { updateTransactions } from 'src/transactions/slice'
 import {
   NetworkId,
   TokenTransaction,
@@ -270,7 +270,7 @@ describe(tagTxsWithProviderInfo, () => {
 
     await expectSaga(
       tagTxsWithProviderInfo,
-      updateTransactions(NetworkId['celo-alfajores'], transactions)
+      updateTransactions({ networkId: NetworkId['celo-alfajores'], transactions })
     )
       .provide([
         [select(providerLogosSelector), mockProviderLogos],
