+++ src/fiatExchanges/saga.ts
@@ -18,7 +18,7 @@ import { AddressRecipient, RecipientType, getDisplayName } from 'src/recipients/
 import { Actions as SendActions } from 'src/send/actions'
 import { TransactionDataInput } from 'src/send/types'
 import { CurrencyTokens, tokensByCurrencySelector } from 'src/tokens/selectors'
-import { Actions as TransactionActions, UpdateTransactionsAction } from 'src/transactions/actions'
+import { UpdateTransactionsPayload, updateTransactions } from 'src/transactions/slice'
 import { Network, TokenTransactionTypeV2 } from 'src/transactions/types'
 import Logger from 'src/utils/Logger'
 import { resolveCurrency } from 'src/utils/currencies'
@@ -129,7 +129,8 @@ export function* fetchTxHashesToProviderMapping() {
   return txHashesToProvider
 }
 
-export function* tagTxsWithProviderInfo({ transactions, networkId }: UpdateTransactionsAction) {
+export function* tagTxsWithProviderInfo(action: UpdateTransactionsPayload) {
+  const { transactions, networkId } = action.payload
   try {
     if (!transactions || !transactions.length || networkIdToNetwork[networkId] !== Network.Celo) {
       return
@@ -171,7 +172,7 @@ export function* watchBidaliPaymentRequests() {
 }
 
 function* watchNewFeedTransactions() {
-  yield* takeEvery(TransactionActions.UPDATE_TRANSACTIONS, safely(tagTxsWithProviderInfo))
+  yield* takeEvery(updateTransactions.type, safely(tagTxsWithProviderInfo))
 }
 
 export function* fiatExchangesSaga() {
